Move require interceptor out of main renderer bundle
