+++ packages/insomnia/config/renderer-node-import-baseline.json
@@ -33,22 +33,6 @@
       "builtin": "path"
     },
     {
-      "importer": "src/script-executor.ts",
-      "builtin": "fs/promises"
-    },
-    {
-      "importer": "src/scripting/require-interceptor.ts",
-      "builtin": "buffer"
-    },
-    {
-      "importer": "src/scripting/require-interceptor.ts",
-      "builtin": "timers"
-    },
-    {
-      "importer": "src/scripting/require-interceptor.ts",
-      "builtin": "util"
-    },
-    {
       "importer": "src/templating/base-extension.ts",
       "builtin": "crypto"
     },
